+++ lib/parser.js
@@ -1,5 +1,6 @@
 var tokenizer = require('./tokenizer');
 var classHandlers = require('./handlers/classHandlers');
+var moduleHandlers = require('./handlers/moduleHandlers');
 
 var multiComment = false;
 var multiLine = false;
@@ -446,49 +447,6 @@ function handleMuchArgs(parseContext)
 function handleLambda(parseContext)
 {
   return 'function' + handleMuchArgs(parseContext);
-}
-
-/**
- * Handles the so construct (thanks @maxogden!):
- *  so <module> [as <name>]
- *
- * Produces:
- *  var name = require('module');
- */
-function handleSo(parseContext)
-{
-  expectToken('so', parseContext);
-
-  var tokens = parseContext.tokens;
-
-  // consume: so
-  tokens.shift();
-
-  var statement = '';
-  var lib = tokens.shift();
-
-  var modName = '';
-
-  // check if it's a named import or not
-  if(tokens.length > 0) {
-    expectToken('as', parseContext);
-    // consume: as
-    tokens.shift();
-    modName = tokens.shift();
-  } else {
-    // so x
-
-    // try to make a simple name
-    var mod = lib;
-    // test for relative module, optional chop extension
-    var m = /^..?\/.*?([\w-]+)(\.\w+)?$/.exec(mod);
-    if (m) {
-        mod = m[1];
-    }
-    modName = mod.replace(/-/g, '_');
-  }
-
-  return `var ${modName} = require('${lib}');\n`;
 }
 
 /**
@@ -1298,7 +1256,7 @@ function parseStatement(parseContext) {
 
   // so require
   if (tokens[0] === 'so') {
-    return handleSo(parseContext);
+    return moduleHandlers.handleSo(parseContext);
   }
 
   // maybe boolean operator
